refactor: rename SaleContext to SaleDbContext and update related references
