--- src/Ambev.DeveloperEvaluation.ORM/SaleContext.cs
@@ -1,93 +1,0 @@
-﻿using Ambev.DeveloperEvaluation.Domain.Entities;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Design;
-using Microsoft.Extensions.Configuration;
-using Ambev.DeveloperEvaluation.Communication.Mediator;
-using Ambev.DeveloperEvaluation.Domain.Repositories;
-using Ambev.DeveloperEvaluation.Messages;
-
-namespace Ambev.DeveloperEvaluation.ORM;
-
-public class SaleContext : DbContext, IUnitOfWork
-{
-    private readonly IMediatorHandler _mediatorHandler;
-
-
-    public SaleContext(DbContextOptions<SaleContext> options, IMediatorHandler mediatorHandler)
-        : base(options)
-    {
-        _mediatorHandler = mediatorHandler;
-    }
-
-    public DbSet<User> Users { get; set; }
-    public DbSet<Sale> Sales { get; set; }
-    public DbSet<SaleItem> SaleItems { get; set; }
-
-    public async Task<bool> CommitAsync(CancellationToken cancellationToken = default)
-    {
-        foreach (var entry in ChangeTracker.Entries()
-                     .Where(entry => entry.Entity.GetType().GetProperty("CreatedAt") != null))
-        {
-            if (entry.State == EntityState.Added)
-            {
-                entry.Property("CreatedAt").CurrentValue = DateTime.UtcNow;
-            }
-
-            if (entry.State == EntityState.Modified)
-            {
-                entry.Property("CreatedAt").IsModified = false;
-            }
-        }
-
-        var success = await base.SaveChangesAsync(cancellationToken) > 0;
-        if (success) await _mediatorHandler.PublishEventsAsync(this);
-
-        return success;
-    }
-
-    protected override void OnModelCreating(ModelBuilder modelBuilder)
-    {
-        foreach (var property in modelBuilder.Model.GetEntityTypes()
-                     .SelectMany(e => e.GetProperties().Where(p => p.ClrType == typeof(string))))
-        {
-            property.SetColumnType("varchar(100)");
-        }
-
-        modelBuilder.Ignore<Event>();
-
-        modelBuilder.ApplyConfigurationsFromAssembly(typeof(SaleContext).Assembly);
-
-        foreach (var relationship in modelBuilder.Model.GetEntityTypes().SelectMany(e => e.GetForeignKeys()))
-        {
-            relationship.DeleteBehavior = DeleteBehavior.ClientSetNull;
-        }
-
-        modelBuilder.HasSequence<int>("SaleSequence").StartsAt(1000).IncrementsBy(1);
-        base.OnModelCreating(modelBuilder);
-    }
-}
-
-public class SaleDbContextFactory : IDesignTimeDbContextFactory<SaleContext>
-{
-    public SaleContext CreateDbContext(string[] args)
-    {
-        var environment = Environment.GetEnvironmentVariable("ASPNETCORE_ENVIRONMENT") ?? "Development";
-
-        IConfigurationRoot configuration = new ConfigurationBuilder()
-            .SetBasePath(Directory.GetCurrentDirectory())
-            .AddJsonFile("appsettings.json")
-            .AddJsonFile($"appsettings.{environment}.json", optional: true)
-            .AddEnvironmentVariables()
-            .Build();
-
-        var builder = new DbContextOptionsBuilder<SaleContext>();
-        var connectionString = configuration.GetConnectionString("DefaultConnection");
-
-        builder.UseNpgsql(
-            connectionString,
-            options => options.MigrationsAssembly(typeof(SaleContext).Assembly.FullName)
-        );
-
-        return new SaleContext(builder.Options, null!); // Pass null for IMediatorHandler in design-time context
-    }
-}
