+++ src/Ambev.DeveloperEvaluation.WebApi/Program.cs
@@ -29,7 +29,7 @@ public class Program
             builder.AddBasicHealthChecks();
             builder.Services.AddSwaggerGen();
 
-            builder.Services.AddDbContext<SaleContext>(options =>
+            builder.Services.AddDbContext<SaleDbContext>(options =>
                 options.UseNpgsql(
                     builder.Configuration.GetConnectionString("DefaultConnection"),
                     b => b.MigrationsAssembly("Ambev.DeveloperEvaluation.ORM")
